requirements.txt
streamlit>=1.32.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
httpx>=0.27.0
trafilatura>=1.8.0
pydantic>=2.6.0
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
pytest>=7.0.0
